+++ 3DCurves/Ellipse.cpp
@@ -1,16 +1,16 @@
 #include "Ellipse.h"
 
-Vector3 Ellipse::getPoint(double t)
+Vector3 Ellipse::getPoint(float t)
 {
 	return {a * std::cos(t), b * std::sin(t), 0.0};
 }
 
-Vector3 Ellipse::getDerivative(double t)
+Vector3 Ellipse::getDerivative(float t)
 {
 	return {-a * std::sin(t), b * std::cos(t), 0.0};
 }
 
-double Ellipse::getRadius()
+float Ellipse::getRadius()
 {
 	return radius;
 }
