+++ 3DCurves/Ellipse.h
@@ -3,21 +3,21 @@
 class Ellipse : public Curve
 {
 public:
-	Ellipse(Vector3 center, double radius, double a, double b, std::string name) : center{ center }, radius{ radius }, name{ name }, a{ a }, b{ b } {}
+	Ellipse(Vector3 center, float radius, float a, float b, std::string name) : center{ center }, radius{ radius }, name{ name }, a{ a }, b{ b } {}
 
 	Vector3 center;
-	double radius;
-	double a;
-	double b;
+	float radius;
+	float a;
+	float b;
 	std::string name;
 
 
-	Vector3 getPoint(double t) override;
+	Vector3 getPoint(float t) override;
 
-	Vector3 getDerivative(double t) override;
+	Vector3 getDerivative(float t) override;
 
 	Vector3 getCenter();
-	double getRadius();
+	float getRadius();
 	std::string getName();
 };
